feature(feedback): lfa feedback
edit feedback from comment

[finishes: #168384229]
